move all prototype code into its own component
